Extracted auth conversion functions into a separate class - AccessTokenAdapter
