+++ itwin-platform-access/imodels-access-frontend/src/IModelsAccessFrontendExports.ts
@@ -2,4 +2,5 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
+export * from "./interface-adapters/AccessTokenAdapter";
 export * from "./FrontendIModelsAccess";
